requirements.txt
python-telegram-bot
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
blockbuster==1.5.24
Bottleneck==1.5.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
cryptography==44.0.3
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
executing==2.2.0
forbiddenfruit==0.1.4
frozenlist==1.6.0
geojson==2.5.0
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-aiplatform==1.94.0
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.16.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-google-vertexai==2.0.24
langchain-openai==0.3.18
langchain-tavily==0.2.0
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-api==0.2.34
langgraph-checkpoint==2.0.26
langgraph-cli==0.2.10
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.2.0
langgraph-sdk==0.1.70
langsmith==0.3.42
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.4.4
mypy==1.16.0
mypy_extensions==1.1.0
mysql-connector-python
nest-asyncio==1.6.0
numexpr==2.10.2
numpy==2.2.6
openai==1.82.1
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
parso==0.8.4
pathspec==0.12.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyowm==3.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.46.2
structlog==25.3.0
tenacity==9.1.2
tiktoken==0.9.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
truststore==0.10.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
validators==0.35.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0

# include other tools/libraries as needed